add phantomjs in docker
There is a problem with the Ubuntu package version of phantomjs, which causes capybara errors.
The version of npm that can be installed with the ubuntu package gives an error when installing phantomjs.
So, install yarn using npm, and install phantomjs using yarn.

see: https://github.com/Medium/phantomjs/issues/707